Added support for compartment reports

+++ plugins/loaders/MeshLoader.h
@@ -45,19 +45,19 @@ public:
      * @param meshContainer structure containing references to triangles and
      *        materials imported from the file, as well as the bounding box of
      *        all meshes.
-     * @param material is set to NO_MATERIAL, materials defined in the source
-     *        file are used. Otherwise, the specified material is assigned to
-     *        all imported meshes.
      * @param meshQuality can be MQ_FAST, MQ_QUALITY or MQ_MAX_QUALITY. Appart
      *        from MQ_FAST, normals are automatically generated is not in the
      *        file.
-     * @param bounds resulting bounding box of the loaded meshes
+     * @param material Default material for the whole mesh. If set to
+     *        NO_MATERIAL, materials from the mesh file are used. Otherwise,
+     *        all meshes are forced to that specific material.
+     * @return true if the file was successfully imported. False otherwise.
      */
     bool importMeshFromFile(
             const std::string& filename,
             MeshContainer& meshContainer,
             MeshQuality meshQuality,
-            int material);
+            size_t material);
 
     /** Exports meshes to a given file
      *
